refactor: structure code following the standard go project layout
