+++ gorvus/internal/commands/root.go
@@ -1,4 +1,4 @@
-package cmd
+package commands
 
 import (
 	"os"
