--- gorvus/main.go
@@ -1,7 +1,0 @@
-package main
-
-import "github.com/FelipeMCassiano/gorvus/gorvus/cmd"
-
-func main() {
-	cmd.Execute()
-}
